Initial commit
This had been a codepen, but I decided to migrate it to github for ease of transferring into other project. It also allowed me to tweak and improve the look slightly from the codepen version.

This is meant to be used as a cutout piece where I will be easily able to grab this clock and place it in other project in varying sizes and looks.