QOL CSS changes making minor tweaks to certain appearances, and making the clock have a max width it can skrink to to prevent wrapping of the time h1
I would like to go back in at some point and have a good long think about how to make the clock face a bit more responsive in size. As it is now it is pretty huge at a minimum 700px across. Plus it would be nice to have a dynamic width over a fixed width.